Added a faster version of day 11
Removed that eval (just used some knowledge about what the input would be, because I definitely wasn't going to probably parse an expression for this), and changed the divisible check to something more sensible. It ended up being significantly faster, but not really in any interesting way.

Keeping the original, because... I don't know. Sentimentality?