Added day 15
This day went poorly...

The first part I had multiple wrong answers, and I'm still not entirely sure how I fixed it (rewrote it in a considerably slower way that should have been the same and then it worked, clearly there was some bug in there).

The second part I messed around far too long on trying to figure out a clever way to do it, before just deciding to rotate the space so that the distances were easier to work with and do the bare minimum optimisation. Turns out that works pretty well.